Completely redesigned everything. See below: First of all, sorry for this insane commit. I forgot to push stuff periodically, so now I gotta push it all at once.
    - Started making Scott's design come to life, so the layout is based on his design.
    - The frame now has a custom shape with rounded corners. This is achieved through
wxFrame::SetShape(), to which I pass a custom-made png image with the needed shape. This means
that it doesn't have a native title bar, so, as of yet, there's no way to close the app
without the use of a task manager (or through debugging in Visual Studio).
    - The left panel uses a wxRichTextCtrl to display the patch notes, since the text needs
to be tightly formatted. For loading the text, it fetches stuff from the website (I actually
will send the pre-formatted needed text to the website, so this control just copies the buffer).
    - Made a custom scrollbar class for wxRichTextCtrl. It's not highly customizable yet, but
I think that I can improve it enough to actually use it in Amadeus-Writer.
    - Made the background of the MainPanel dynamic: It offsets the image according to the
mouse movement. It pre-positions the image so that the maximum offset is never bigger than
the image height or width.
    - Actually used a wxSFShapeCanvas for the main panel because I needed to make the buttons
more customized than possible with the native controls. This is done thorugh a TransparentButton
class, which inherits from wxSFRoundRectShape. I still haven't figured out how I'm gonna deal
with the positioning of the needed bitmaps inside of it, only text.
    - Added a HyperlinkPanel class, which displays a bitmap (scaled to the panel's size) and
of course opens the browser with the provided url on click. It also has a pseudo hover effect,
which draws a semi-transparent black layer on top of the panel if the user isn't hovering it.
    - Started drawing pretty stuff in the PatchNotesWindow through clever use of the
wxDC gradient functions.

There're probably some things that I'm forgetting, but as far as I can tell this covers most
of it. I'll of course start commiting more frequently now.